Add files via upload
The objective of this project is to conduct an in-depth analysis of the terrorism dataset through Exploratory Data Analysis (EDA). Our aim is to discover valuable insights and recurring patterns within the data. Through this analysis, we intend to unveil trends, pinpoint regions with heightened risk, gain insights into the methods used in attacks, and potentially shed light on the underlying factors that influence terrorist activities.